refactor(tools): move DeferredFunctionToolSchema to tools.final
Relocate DeferredFunctionToolSchema from ag2.tools.builtin.tool_search
to ag2.tools.final.function_tool to improve organization.
