+++ ag2/config/openai/mappers.py
@@ -37,9 +37,9 @@ from ag2.tools.builtin.shell import (
     ShellToolSchema,
 )
 from ag2.tools.builtin.skills import SkillsToolSchema
-from ag2.tools.builtin.tool_search import DeferredFunctionToolSchema, ToolSearchToolSchema
+from ag2.tools.builtin.tool_search import ToolSearchToolSchema
 from ag2.tools.builtin.web_search import WebSearchToolSchema
-from ag2.tools.final import FunctionToolSchema
+from ag2.tools.final import DeferredFunctionToolSchema, FunctionToolSchema
 from ag2.tools.schemas import ToolSchema
